Move Configuration Values into a Config Type
This adds a `Config` type to the new `HsNotifications.Config` module &
modifies the `run` function to take a Config. A default instance has
been supplied based off the Molokai vim colorscheme. This allows users
to customize their notifications without editing source code & opens the
way for an INI-style config file.

The config is passed through to most functions, which is pretty ugly.
What should eventually be done is creating an App type that composes
ReaderT & StateT for the Config & AppState. That way neither of those
value have to be threaded through functions that don't need them(but
call functions that do).