Add Actions/Buttons to Notifications
Extend the Notification type to include the list of possible actions
sent over dbus. Modify the notificaiton builder so it creates buttons
for any non-default actions. Default actions are currently ignored. Add
a emitActionInvoked function to trigger the action via dbus when the
button is clicked.

Add the `actions` capability to the getCapabilities dbus route - this
indicates we support buttons/actions & is required to get firefox to use
the notification server instead of it's own internal implementation of
notifications.

Refactor the appRemovalQueue into an appEventQueue that contains both
removal & invoke-action events for the dbus-server to process.