Render Images From Notification Data
Parse & render the `image-data` hint from notifications. Add new config
options for limiting the width & height of rendered images.

Add a new `ImageData` type along with `IsVariant` instances so we can
parse it out of the `hints` map if present when we receive a
notification.

If a notification includes an `ImageData`, scale it down if necessary &
render it in a separate column to the left of the title & body text.
Switch from the simple `container_add` element insertion function to the
more thorough `grid_attach` function. Tweak the grid, title, & label
elements to ensure everything is properly aligned & spaced when
rendering images.